src/ContentProcessor/pyproject.toml
[project]
name = "contentprocessor"
version = "0.1.0"
description = "Content Process Gold Standard Solution Accelerator - Content Processing Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-appconfiguration>=1.7.1",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.1",
    "azure-storage-queue>=12.12.0",
    "certifi>=2024.12.14",
    "charset-normalizer>=3.4.1",
    "openai==1.65.5",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "poppler-utils>=0.1.0",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "pymongo>=4.11.2",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.9.0",
]


[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "pydantic>=2.10.5",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.1",
]

[tool.pytest.ini_options]
minversion = "8.3.4"
addopts = "--maxfail=1"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.poetry.scripts]
start-app = "src.main:main"
